More maze improvements
There's too many small spaces I want to have in the maze and it doesn't make sense to just have everything be trees, so I will try using hedges for some sections instead as the "walls". It'll mostly just be a tedious ask at this point, there's nothing we need for the maze coding/asset wise we don't have already.